# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Luna Athletics API",
        "description": "API para gestión de productos y categorías de Luna Athletics",
        "contact": {
            "email": "admin@lunaathletics.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/prendas": {
            "get": {
                "tags": [
                    "Prendas"
                ],
                "summary": "Obtener todas las prendas",
                "operationId": "9c6d236a58c12332aeabaca44186250d",
                "responses": {
                    "200": {
                        "description": "Lista de prendas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PrendaCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Prendas"
                ],
                "summary": "Crear una nueva prenda",
                "operationId": "051cb77a9c07540d4d107381ea3ebfd9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePrendaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Prenda creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Prenda"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/prendas/{id}": {
            "get": {
                "tags": [
                    "Prendas"
                ],
                "summary": "Obtener una prenda específica",
                "operationId": "794115c2eb083eecc1c9f620d7c41a3f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Prenda encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Prenda"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Prenda no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Prendas"
                ],
                "summary": "Actualizar una prenda",
                "operationId": "4600dbc2491704c5ac544b5e31991c30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePrendaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Prenda actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Prenda"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Prendas"
                ],
                "summary": "Eliminar una prenda",
                "operationId": "46eedbf183417fffce2c582298473707",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Prenda eliminada exitosamente"
                    },
                    "404": {
                        "description": "Prenda no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Prenda": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tipo": {
                        "type": "string",
                        "example": "prenda"
                    },
                    "atributos": {
                        "properties": {
                            "titulo": {
                                "type": "string",
                                "example": "Camiseta Deportiva"
                            },
                            "descripcion": {
                                "type": "string",
                                "example": "Descripción de la prenda"
                            },
                            "imagen": {
                                "type": "string",
                                "example": "prendas/imagen.jpg"
                            },
                            "categoria": {
                                "type": "string",
                                "example": "Ropa Deportiva"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "PrendaCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Prenda"
                        }
                    }
                },
                "type": "object"
            },
            "StorePrendaRequest": {
                "required": [
                    "categoria_id",
                    "titulo",
                    "descripcion",
                    "imagen"
                ],
                "properties": {
                    "categoria_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "titulo": {
                        "type": "string",
                        "example": "Prenda 1"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "descripcion de la prenda"
                    },
                    "imagen": {
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearer_token": {
                "type": "apiKey",
                "description": "Sanctum bearer token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}